+++ shared/converter/index.ts
@@ -4,4 +4,6 @@ export { convertJsonToMarkdown } from './jsonToMarkdown';
 export type { SdocMeta as MdSdocMeta } from './jsonToMarkdown';
 export { convertJsonToAdoc } from './jsonToAdoc';
 export type { SdocMeta as AdocSdocMeta } from './jsonToAdoc';
+export { convertJsonToSlides } from './jsonToSlides';
+export type { SdocMeta as SlideSdocMeta } from './jsonToSlides';
 export { convertMarkdownToJson } from './markdownToJson';